Address feedback from @joehan and go/atad-cli-api-review
