+++ src/appdistribution/client.ts
@@ -6,113 +6,16 @@ import { Distribution } from "./distribution";
 import { FirebaseError } from "../error";
 import { Client } from "../apiv2";
 import { appDistributionOrigin } from "../api";
-
-/**
- * Helper interface for an app that is provisioned with App Distribution
- */
-export interface AabInfo {
-  name: string;
-  integrationState: IntegrationState;
-  testCertificate: TestCertificate | null;
-}
-
-export interface TestCertificate {
-  hashSha1: string;
-  hashSha256: string;
-  hashMd5: string;
-}
-
-/** Enum representing the App Bundles state for the App */
-export enum IntegrationState {
-  AAB_INTEGRATION_STATE_UNSPECIFIED = "AAB_INTEGRATION_STATE_UNSPECIFIED",
-  INTEGRATED = "INTEGRATED",
-  PLAY_ACCOUNT_NOT_LINKED = "PLAY_ACCOUNT_NOT_LINKED",
-  NO_APP_WITH_GIVEN_BUNDLE_ID_IN_PLAY_ACCOUNT = "NO_APP_WITH_GIVEN_BUNDLE_ID_IN_PLAY_ACCOUNT",
-  APP_NOT_PUBLISHED = "APP_NOT_PUBLISHED",
-  AAB_STATE_UNAVAILABLE = "AAB_STATE_UNAVAILABLE",
-  PLAY_IAS_TERMS_NOT_ACCEPTED = "PLAY_IAS_TERMS_NOT_ACCEPTED",
-}
-
-export enum UploadReleaseResult {
-  UPLOAD_RELEASE_RESULT_UNSPECIFIED = "UPLOAD_RELEASE_RESULT_UNSPECIFIED",
-  RELEASE_CREATED = "RELEASE_CREATED",
-  RELEASE_UPDATED = "RELEASE_UPDATED",
-  RELEASE_UNMODIFIED = "RELEASE_UNMODIFIED",
-}
-
-export interface Release {
-  name: string;
-  releaseNotes: ReleaseNotes;
-  displayVersion: string;
-  buildVersion: string;
-  createTime: Date;
-  firebaseConsoleUri: string;
-  testingUri: string;
-  binaryDownloadUri: string;
-}
-
-export interface ReleaseNotes {
-  text: string;
-}
-
-export interface UploadReleaseResponse {
-  result: UploadReleaseResult;
-  release: Release;
-}
-
-export interface BatchRemoveTestersResponse {
-  emails: string[];
-}
-
-export interface Group {
-  name: string;
-  displayName: string;
-  testerCount?: number;
-  releaseCount?: number;
-  inviteLinkCount?: number;
-}
-
-export interface CreateReleaseTestRequest {
-  releaseTest: ReleaseTest;
-}
-
-export enum TestState {
-  IN_PROGRESS = "IN_PROGRESS",
-  PASSED = "PASSED",
-  FAILED = "FAILED",
-  INCONCLUSIVE = "INCONCLUSIVE",
-}
-
-export interface TestDevice {
-  model: string;
-  version: string;
-  locale: string;
-  orientation: string;
-}
-
-export interface DeviceExecution {
-  device: TestDevice;
-  state?: TestState;
-  failedReason?: string;
-  inconclusiveReason?: string;
-}
-
-export interface FieldHints {
-  usernameResourceName?: string;
-  passwordResourceName?: string;
-}
-
-export interface LoginCredential {
-  username?: string;
-  password?: string;
-  fieldHints?: FieldHints;
-}
-
-export interface ReleaseTest {
-  name?: string;
-  deviceExecutions: DeviceExecution[];
-  loginCredential?: LoginCredential;
-}
+import {
+  AabInfo,
+  BatchRemoveTestersResponse,
+  DeviceExecution,
+  Group,
+  LoginCredential,
+  ReleaseTest,
+  TestDevice,
+  UploadReleaseResponse,
+} from "./types";
 
 /**
  * Makes RPCs to the App Distribution server backend.
@@ -321,7 +224,6 @@ export class AppDistributionClient {
           loginCredential,
         },
       });
-      utils.logSuccess(`Release test created successfully`);
       return response.body;
     } catch (err: any) {
       throw new FirebaseError(`Failed to create release test ${err}`);
